# docs/swagger.yaml
basePath: /
definitions:
  doc.CPDetail:
    properties:
      keterangan:
        type: string
      name:
        type: string
      ok:
        type: boolean
    type: object
  doc.DeviceDocForm:
    properties:
      checkpoint:
        items:
          $ref: '#/definitions/doc.CPDetail'
        type: array
      device_id:
        type: string
    type: object
  doc.UpdateDeviceDocForm:
    properties:
      checkpoint:
        items:
          $ref: '#/definitions/doc.CPDetail'
        type: array
    type: object
...
  handler.printerForm:
    properties:
      departemen:
        type: string
      nama:
        type: string
      no_seri:
        type: string
      tipe_printer:
        type: string
    type: object
  handler.teleponForm:
    properties:
      departemen:
        type: string
      ext:
        type: string
      lokasi:
        type: string
      merk:
        type: string
      tipe:
        type: string
      user:
        type: string
    type: object
...
  handler.updateCheckpointForm:
    properties:
      checkpoint:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  description: Sistem Pencatatan Maintenance IT Backend API
  title: Sistem Pencatatan Maintenance IT Backend
paths:
...
  /api/checkpoint/printer:
    get:
      operationId: get-printer-checkpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get printer checkpoint
      tags:
      - Checkpoint
    put:
      consumes:
      - application/json
      operationId: update-printer-checkpoint
      parameters:
      - description: Checkpoint
        in: body
        name: checkpoint
        required: true
        schema:
          $ref: '#/definitions/handler.updateCheckpointForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update printer checkpoint
      tags:
      - Checkpoint
  /api/checkpoint/telepon:
    get:
      operationId: get-telepon-checkpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get telepon checkpoint
      tags:
      - Checkpoint
    put:
      consumes:
      - application/json
      operationId: update-telepon-checkpoint
      parameters:
      - description: Checkpoint
        in: body
        name: checkpoint
        required: true
        schema:
          $ref: '#/definitions/handler.updateCheckpointForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update telepon checkpoint
      tags:
      - Checkpoint
...
  /api/doc/printer:
    post:
      operationId: create-printer-document
      parameters:
      - description: Printer Document Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.DeviceDocForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create printer document
      tags:
      - Doc Printer
  /api/doc/printer/{id}:
    delete:
      operationId: delete-printer-document-by-id
      parameters:
      - description: Printer Document ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete printer document by ID
      tags:
      - Doc Printer
    get:
      operationId: get-printer-document-by-id
      parameters:
      - description: Printer Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get printer document by ID
      tags:
      - Doc Printer
    put:
      operationId: update-printer-document-by-id
      parameters:
      - description: Printer Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Printer Document Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.UpdateDeviceDocForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update printer document by ID
      tags:
      - Doc Printer
  /api/doc/printers:
    get:
      operationId: get-all-printer-documents
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get all printer documents
      tags:
      - Doc Printer
  /api/doc/telepon:
    post:
      consumes:
      - application/json
      operationId: create-new-telepon-document
      parameters:
      - description: Telepon Document Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.DeviceDocForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create new telepon document
      tags:
      - Doc Telepon
  /api/doc/telepon/{id}:
    delete:
      operationId: delete-telepon-document-by-id
      parameters:
      - description: Telepon Document ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete telepon document by ID
      tags:
      - Doc Telepon
    get:
      operationId: get-telepon-document-by-id
      parameters:
      - description: Telepon Document ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get telepon document by ID
      tags:
      - Doc Telepon
    put:
      operationId: update-telepon-document-by-id
      parameters:
      - description: Telepon Document ID
        in: path
        name: id
        required: true
        type: string
      - description: Telepon Document Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/doc.UpdateDeviceDocForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update telepon document by ID
      tags:
      - Doc Telepon
  /api/doc/telepons:
    get:
      operationId: get-all-telepons-documents
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get all telepons documents
      tags:
      - Doc Telepon
...
  /api/printer:
    post:
      operationId: create-new-printer
      parameters:
      - description: Printer Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.printerForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create new printer
      tags:
      - Device Printer
  /api/printer/{id}:
    delete:
      operationId: delete-printer-by-id
      parameters:
      - description: Printer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete printer by id
      tags:
      - Device Printer
    get:
      operationId: get-printer-by-id
      parameters:
      - description: Printer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get printer by id
      tags:
      - Device Printer
    put:
      operationId: update-printer-by-id
      parameters:
      - description: Printer ID
        in: path
        name: id
        required: true
        type: string
      - description: Printer Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.printerForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update printer by id
      tags:
      - Device Printer
  /api/printers:
    get:
      operationId: get-all-printers
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get all printers
      tags:
      - Device Printer
  /api/telepon:
    post:
      operationId: create-new-telepon
      parameters:
      - description: Telepon Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.teleponForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create new telepon
      tags:
      - Device Telepon
  /api/telepon/{id}:
    delete:
      operationId: delete-telepon-by-id
      parameters:
      - description: Telepon ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete telepon by id
      tags:
      - Device Telepon
    get:
      operationId: get-telepon-by-id
      parameters:
      - description: Telepon ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get telepon by id
      tags:
      - Device Telepon
    put:
      operationId: update-telepon-by-id
      parameters:
      - description: Telepon ID
        in: path
        name: id
        required: true
        type: string
      - description: Telepon Form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.teleponForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update telepon by id
      tags:
      - Device Telepon
  /api/telepons:
    get:
      operationId: get-all-telepons
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get all telepons
      tags:
      - Device Telepon
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"